Split options into ES6 modules
