+++ src/manifest.json
@@ -23,7 +23,7 @@
   },
 
   "options_ui": {
-    "page": "options.html",
+    "page": "options/options.html",
     "browser_style": true
   },
 
@@ -58,7 +58,7 @@
   "applications": {
     "gecko": {
       "id": "offline-qr-code@rugk.github.io",
-      "strict_min_version": "57.0a1"
+      "strict_min_version": "60.0a1"
     }
   }
 }
