+++ src/options/options.html
@@ -7,8 +7,8 @@
 		<link rel="stylesheet" href="options.css">
 
 		<script src="common/lib/lodash/lodash.js"></script>
-		<script defer src="common/common.js"></script>
-		<script defer src="options.js"></script>
+		<script defer src="common/common.js" type="module"></script>
+		<script defer src="options.js" type="module"></script>
 	</head>
 
 	<body>
